Fix x86 warning

+++ source/script.cpp
@@ -6759,7 +6759,7 @@ LPTSTR Script::FindLibraryFile(LPTSTR aFuncName, size_t aFuncNameLength, bool aI
 	// The legacy behaviour for #Include <A_B> is that all Libs are searched for A_B.ahk before
 	// searching for A.ahk, which means that A_B.ahk takes precedence over A.ahk even if A.ahk
 	// is defined in the local Lib and A_B.ahk is not.
-	for (i = 0; i < aFuncNameLength; ++i)
+	for (i = 0; i < (int)aFuncNameLength; ++i)
 		if (aFuncName[i] == '_')
 			return FindLibraryFile(aFuncName, i);
 	return nullptr;
